refactor(sync): extract sync-keys module to break circular imports
Move sync key helpers out of sync.ts so store and component tests can import
them without pulling the full sync pipeline.
