+++ src/lib/local/data/sync.ts
@@ -11,56 +11,9 @@ import { getDB } from "./pgliteDB";
 import { fetchJsonWithRetry, SYNC_CHECK_FETCH_OPTIONS } from "./fetch-json";
 import { syncToastStore } from "@lib/store.svelte";
 import type { Results } from "@electric-sql/pglite";
+import { getSyncKeysFromLs } from "./sync-keys";
 
-export function getSyncKeysFromLs(): {
-  [key: string]: string;
-} | null {
-  const lsStore = localStorage.getItem("sync-key");
-
-  if (lsStore === null) {
-    localStorage.setItem(
-      "sync-key",
-      `{
-      "buildings": "",
-      "colleges": "",
-      "divisions": "",
-      "rooms": "",
-      "dorms": "",
-      "classes": "",
-      "final_exams": "",
-      "events": ""
-    }`,
-    );
-    return null;
-  }
-
-  try {
-    const keys = JSON.parse(lsStore);
-    if (typeof keys === "object" && "buildings" in keys) {
-      return keys as {
-        [key: string]: string;
-      };
-    } else {
-      return null;
-    }
-  } catch {
-    console.error("Error: the sync keys in the localStorage was corrupted");
-    localStorage.setItem(
-      "sync-key",
-      `{
-      "buildings": "",
-      "colleges": "",
-      "divisions": "",
-      "rooms": "",
-      "dorms": "",
-      "classes": "",
-      "final_exams": "",
-      "events": ""
-    }`,
-    );
-    return null;
-  }
-}
+export { getSyncKeysFromLs };
 
 async function localCacheIsEmpty(tableName: string): Promise<boolean> {
   try {
